# container_imgs/city-tags-api/docs/swagger.yaml
basePath: /api/v0
definitions:
  api_errors.ClientErr:
    properties:
      code:
        type: integer
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  server.CityData:
    properties:
      city_id:
        type: integer
      city_name:
        type: string
      continent:
        type: string
      country_3_code:
        type: string
    type: object
  server.TagsData:
    properties:
      air_quality:
        type: string
      city_id:
        type: integer
      city_size:
        type: string
      cloud_coverage:
        type: string
      daylight_hours:
        type: string
      humidity:
        type: string
      precipitation:
        type: string
      temperature:
        type: string
    type: object
host: city-tags-api.com
info:
  contact:
    email: a.jaenrev@gmail.com
    name: City Tags API Support
  description: This is an API that makes available different tags for worlwide cities
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: City Tags API
  version: 0.0.7
paths:
  /v0/cities:
    get:
      consumes:
      - application/json
      description: Get cities with pagination
      parameters:
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Limit for pagination
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.CityData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_errors.ClientErr'
      summary: Get cities
  /v0/cities/{cityId}:
    get:
      consumes:
      - application/json
      description: Get city information by providing a specific city id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.CityData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_errors.ClientErr'
      summary: Get city by city id
  /v0/cities/{cityId}/tags:
    get:
      consumes:
      - application/json
      description: Get tags information by providing a specific city id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.TagsData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_errors.ClientErr'
      summary: Get city tags by city id
securityDefinitions:
  ApiKeyAuth:
    description: Authorization to access the API endpoints
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"